Packager#cookbook_files is now Packager#packaging_slip
The purpose of the packaging slip is to map from physical file paths to
tarball file paths. Such a construct makes Packager#tar effectively a
static method once again.

The packaging_slip method could be about half as long as it is. However,
I think using the style of successive maps highlights the
transformations required to turn an absolute path to a file into a
relative path for the tarball.